extract Quiz page helpers
I'm hoping, in the future, we can extract similar helpers from other
tests as well.
